- Updated `config.py` to use `DoubleType` for latitude and longitude fields. - Improved robustness in distance calculations in `simulation.py` with clipping.. - Removed outdated `data_quality.py` and replaced it with `test_data_quality.py` - Created `test_etl.py` to validate the ETL process with various scenarios. - Updated `test_simulation.py` to reflect accurate geographical coordinates and ensure simulation logic correctness.

+++ src/config.py
@@ -1,4 +1,4 @@
-from pyspark.sql.types import StructType, StructField, StringType, IntegerType, FloatType
+from pyspark.sql.types import StructType, StructField, StringType, IntegerType, FloatType, DoubleType
 
 # --- PHYSICS CONSTANTS ---
 METRO_POPULATION_THRESHOLD = 200000 
@@ -37,8 +37,8 @@ SCHEMA = StructType([
     StructField("name", StringType(), True),
     StructField("asciiname", StringType(), True),
     StructField("alternatenames", StringType(), True),
-    StructField("latitude", FloatType(), True),
-    StructField("longitude", FloatType(), True),
+    StructField("latitude", DoubleType(), True),
+    StructField("longitude", DoubleType(), True),
     StructField("feature_class", StringType(), True),
     StructField("feature_code", StringType(), True),
     StructField("country_code", StringType(), True),
